# swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Codecademy e-Commerce REST API",
      "description": "Defines the backend logic relating to the <strong>'registration'</strong>, <strong>'customers'</strong>, <strong>'addresses'</strong>, <strong>'orders'</strong>, <strong>'products'</strong>, <strong>'cart'</strong>, and <strong>'login'</strong> endpoints of an e-commerce REST API.  <code>GET</code> operations can be performed directly on the <strong>HEROKU</strong> database; <code>POST</code>/<code>PUT</code>/<code>DELETE</code> operations can be performed on the <strong>SWAGGER HUB</strong> <em>virtual mocking server</em> so that the <strong>HEROKU</strong> database is not compromised.",
      "version": "0.1"
    },
    "servers": [
      {
        "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
        "description": "Heroku App"
      },
      {
        "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
        "description": "SwaggerHub API Auto Mocking"
      }
    ],
    "paths": {
...
      "/api/orders/{customerId}": {
        "get": {
          "tags": [
            "Orders"
          ],
          "summary": "Return the order history for a particular customer",
          "description": "Returns the order history of the customer whose id is 'customerId'",
          "parameters": [
            {
              "name": "customerId",
              "in": "path",
              "description": "The user ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Returns an array of JSON objects representing the order history of the specified customer",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Order"
                    }
                  },
                  "example": [
                    {
                      "order_id": 3,
                      "date_of_purchase": "2022-02-25T23:52:32.080Z",
                      "cart": [
                        {
                          "product_id": 3,
                          "customer_id": 12,
                          "product_name": "Opportunity Rover - 2004-01-26 (Sol 1)",
                          "product_description": "Photo taken by the Opportunity Rover's Panoramic Camera on 2004-01-26 (expedition sol 1)",
                          "product_image_url": "https://mars.nasa.gov/mer/gallery/all/1/p/001/1P128287426EFF0000P2303L2M1-BR.JPG",
                          "product_price": 600,
                          "product_quantity": 9,
                          "cumulative_product_price": 5400,
                          "cart_total": 5400
                        }
                      ]
                    },
                    {
                      "order_id": 2,
                      "date_of_purchase": "2022-02-23T21:04:14.510Z",
                      "cart": [
                        {
                          "product_id": 5,
                          "customer_id": 12,
                          "product_name": "Opportunity Rover - 2004-01-26 (Sol 1)",
                          "product_description": "Photo taken by the Opportunity Rover's Rear Hazard Avoidance Camera on 2004-01-26 (expedition sol 1)",
                          "product_image_url": "https://mars.nasa.gov/mer/gallery/all/1/r/001/1R128285184EDN0000P1002L0M1-BR.JPG",
                          "product_price": 900,
                          "product_quantity": 5,
                          "cumulative_product_price": 4500,
                          "cart_total": 7300
                        },
                        {
                          "product_id": 4,
                          "customer_id": 12,
                          "product_name": "Opportunity Rover - 2009-09-11 (Sol 2001)",
                          "product_description": "Photo taken by the Opportunity Rover's Rear Hazard Avoidance Camera on 2009-09-11 (expedition sol 2001)",
                          "product_image_url": "https://mars.nasa.gov/mer/gallery/all/1/r/2001/1R305828935EFFA5FWP1312L0M1-BR.JPG",
                          "product_price": 700,
                          "product_quantity": 4,
                          "cumulative_product_price": 2800,
                          "cart_total": 7300
                        }
                      ]
                    }
                  ]
                }
              }
            },
            "404": {
              "description": "No orders belonging to the specified customer were found in the database",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "No orders belonging to the specified customer were found in the database."
                  }
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
              "description": "Heroku App"
            },
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "servers": [
          {
            "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
            "description": "Heroku App"
          },
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
            "description": "SwaggerHub API Auto Mocking"
          }
        ]
      },
      "/api/orders/{customerId}/{orderId}": {
        "get": {
          "tags": [
            "Orders"
          ],
          "summary": "Return the specified order details for the specified customer",
          "description": "Returns the order details of the order whose id is 'orderId', made by the customer whose id is 'customerId'",
          "parameters": [
            {
              "name": "customerId",
              "in": "path",
              "description": "The user ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            },
            {
              "name": "orderId",
              "in": "path",
              "description": "The order ID",
              "required": true,
              "style": "simple",
              "explode": false,
              "schema": {
                "minimum": 1,
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Returns a single-element array containing the specified order details for the specified customer",
              "content": {
                "application/json; charset=utf-8": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Order"
                    }
                  },
                  "example": [
                    {
                      "order_id": 3,
                      "date_of_purchase": "2022-02-25T23:52:32.080Z",
                      "cart": [
                        {
                          "product_id": 3,
                          "customer_id": 12,
                          "product_name": "Opportunity Rover - 2004-01-26 (Sol 1)",
                          "product_description": "Photo taken by the Opportunity Rover's Panoramic Camera on 2004-01-26 (expedition sol 1)",
                          "product_image_url": "https://mars.nasa.gov/mer/gallery/all/1/p/001/1P128287426EFF0000P2303L2M1-BR.JPG",
                          "product_price": 600,
                          "product_quantity": 9,
                          "cumulative_product_price": 5400,
                          "cart_total": 5400
                        }
                      ]
                    }
                  ]
                }
              }
            },
            "404": {
              "description": "No association of the specified 'customerId' and 'orderId' exists in the database",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "No association of the specified customer ID and order ID exists in the database."
                  }
                }
              }
            },
            "500": {
              "description": "Returns a description of the error {err} that was unexpectedly encountered",
              "content": {
                "text/html; charset=utf-8": {
                  "schema": {
                    "type": "string",
                    "example": "The unexpected value of {err.message}."
                  }
                }
              }
            }
          },
          "servers": [
            {
              "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
              "description": "Heroku App"
            },
            {
              "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
              "description": "SwaggerHub API Auto Mocking"
            }
          ]
        },
        "servers": [
          {
            "url": "https://codecademy-e-commerce-rest-api.herokuapp.com",
            "description": "Heroku App"
          },
          {
            "url": "https://virtserver.swaggerhub.com/chris-larham-1983/e-Commerce_Registration_Customers_Addresses_Orders/0.1",
            "description": "SwaggerHub API Auto Mocking"
          }
        ]
      },
...
    },
    "components": {
      "schemas": {
...
        "Order": {
          "type": "object",
          "properties": {
            "order_id": {
              "type": "integer",
              "description": "The order ID"
            },
            "date_of_purchase": {
              "type": "string",
              "description": "The timestamp associated with the order",
              "format": "date-time"
            },
            "cart": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Order_cart"
              }
            }
          }
        },
...
        "Order_cart": {
          "type": "object",
          "properties": {
            "product_id": {
              "type": "integer",
              "description": "The product ID"
            },
            "customer_id": {
              "type": "integer",
              "description": "The customer ID"
            },
            "product_name": {
              "type": "string",
              "description": "The product name"
            },
            "product_description": {
              "type": "string",
              "description": "The product description"
            },
            "product_image_url": {
              "type": "string",
              "description": "The URL of the image that represents the product"
            },
            "product_price": {
              "type": "integer",
              "description": "A representation of the product price in pence"
            },
            "product_quantity": {
              "type": "integer",
              "description": "The number of items of this product that was purchased"
            },
            "cumulative_product_price": {
              "type": "integer",
              "description": "The result of multiplying the product price by the product quantity"
            },
            "cart_total": {
              "type": "integer",
              "description": "The sum total of all items in this order"
            }
          }
        }
      }
    }
  }